chore: Rename OpenAI client to LlamaCppClient and update imports
