+++ backend/api/deps.py
@@ -5,7 +5,7 @@ Defines dependencies used by the endpoints.
 from typing import Annotated, Generator
 
 import state
-from bot.client.openai_client import LlamaCppClient
+from bot.client.llamacpp_client import LlamaCppClient
 from bot.conversation.chat_history import ChatHistory
 from bot.memory.vector_database.chroma import Chroma
 from chat_history import chat_history
