+++ backend/state.py
@@ -4,7 +4,7 @@ Holds singleton instances that are initialized during app startup.
 """
 
 
-from bot.client.openai_client import LlamaCppClient
+from bot.client.llamacpp_client import LlamaCppClient
 from bot.memory.vector_database.chroma import Chroma
 from sqlalchemy import Engine
