+++ chatbot/bot/conversation/conversation_handler.py
@@ -4,7 +4,7 @@ from typing import Any
 from entities.document import Document
 from helpers.log import get_logger
 
-from bot.client.openai_client import LlamaCppClient
+from bot.client.llamacpp_client import LlamaCppClient
 from bot.conversation.chat_history import ChatHistory
 from bot.conversation.ctx_strategy import BaseSynthesisStrategy
