+++ chatbot/bot/conversation/ctx_strategy.py
@@ -5,7 +5,7 @@ from typing import Any
 from entities.document import Document
 from helpers.log import get_logger
 
-from bot.client.openai_client import LlamaCppClient
+from bot.client.llamacpp_client import LlamaCppClient
 
 logger = get_logger(__name__)
